Move api/server/router/debug to daemon/server/router/debug
